pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imagerenamer"
dynamic = ["version"]
description = "Rename image files based on their creation date from EXIF metadata"
readme = "README.md"
authors = [
    {name = "Lars van der Niet", email = "lvdnbusiness@icloud.com"}
]
license = {text = "MIT"}
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow>=9.0.0",
    "PyQt6>=6.4.0",
]

[project.urls]
"Homepage" = "https://github.com/larsniet/image-renamer"
"Bug Tracker" = "https://github.com/larsniet/image-renamer/issues"

[project.scripts]
imagerenamer = "imagerenamer.cli:main"
imagerenamer-gui = "imagerenamer.gui:main" 